+++ src/main/java/com/newhead/barablah/modules/barablahmembercomment/base/protocol/AbstractBarablahMemberCommentQueryListRequest.java
@@ -6,7 +6,7 @@ import java.math.BigDecimal;
 
 /**
  * 
- * @generated 2018年03月24日 01:24:55
+ * @generated 2018年03月25日 10:55:21
  */
 @Data
 public abstract class AbstractBarablahMemberCommentQueryListRequest  {
@@ -14,7 +14,7 @@ public abstract class AbstractBarablahMemberCommentQueryListRequest  {
      *
      * 班级
      */
-     private Integer classId;
+     private Long classId;
 
     /**
      *
